Remove forking and update image check condition

+++ src/zwift.sh
@@ -299,16 +299,14 @@ fi
 
 # Clean previous container images (if any)
 if [[ ! $DONT_CLEAN ]] && [[ ! $DONT_PULL ]]; then
-    readarray -t images < <($CONTAINER_TOOL images --filter "reference=${IMAGE/docker.io\/}" --filter "before=${IMAGE/docker.io\/}:$VERSION" --format '{{.ID}}')
-    if [ "${#images[@]}" != "0" ]; then
+    readarray -t images < <($CONTAINER_TOOL images --filter "reference=${IMAGE#docker.io/}" --filter "before=${IMAGE#docker.io/}:$VERSION" --format '{{.ID}}')
+    if [ "${#images[@]}" -gt 0 ] && [ -n "${images[0]}" ]; then
         msgbox info "Cleaning up previous container images"
-        (
-            if $CONTAINER_TOOL image rm "${images[@]}"; then
-                msgbox ok "Previous container images have been deleted"
-            else
-                msgbox error "Failed to clean up previous container images"
-            fi
-        ) &
+        if $CONTAINER_TOOL image rm "${images[@]}"; then
+            msgbox ok "Previous container images have been deleted"
+        else
+            msgbox warning "Failed to clean up previous container images"
+        fi
     fi
 fi
